Cleanup Coverity Scan Bug
Coverity scan is getting mad about our use of new / delete
in our unit tests, even though we mimic the standard library
(it's a false positive), so this patch moves this code to a
system header to prevent the scanner from see it. It also
will provide a generic way to do this for other unit tests
in the future.

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/193
